🏗️ Fix #1237: アーキテクチャ統合完了 - ファイルサイズ最適化達成
主な変更:
- unified_api.py ファイルサイズ最適化: 565行→430行 (135行削減)
- API便利関数を api_utils.py に分離 (107行)
- 要素カウンター機能を element_counter.py に分離 (82行)
- コード品質向上: Black準拠フォーマット適用
- 機能性維持: 統合API動作確認済み

成果:
- Manager統合システム正常動作確認
- テストカバレッジ 43.21% 維持
- 後方互換性完全保持
